Add implementation of report generation

+++ src/Report.cs
@@ -9,4 +9,4 @@ public record Report(Instant ExtractedAt, ImmutableList<ReportLine> Lines);
 /// </summary>
 /// <param name="LocalTime">The local time marking the start of the hour period the line applies to.</param>
 /// <param name="Volume">The aggregated trade positions for the specified hour.</param>
-public record ReportLine(LocalTime LocalTime, int Volume);
+public record ReportLine(LocalTime LocalTime, double Volume);
